requirements.txt
python-dotenv>=1.2.1
langchain>=0.3.27
faiss-cpu>=1.13.2
pypdf>=3.0.0
tiktoken>=0.3.3
docx2txt>=0.1.0
langchain-community>=0.3.27
langchain-classic>=0.3.27
langchain-google-genai>=0.3.27